requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
langchain
langchain-openai
python-dotenv
openai
pymongo